moves around types since its a lot for one file
