+++ lib/types/ApiError.ts
@@ -1,0 +1,18 @@
+export interface ApiError {
+  /**
+   * The Id of the specific error occurrence.
+   * Please include this information when contacting Smartsheet support.
+   */
+  refId: string;
+
+  /**
+   * Custom error code from Smartsheet. See the complete Error Code List.
+   * https://smartsheet.redoc.ly/#section/Error-Code-List
+   */
+  errorCode: number;
+
+  /**
+   * Descriptive error message.
+   */
+  message: string;
+}
