+++ lib/types/ApiUrls.ts
@@ -1,0 +1,20 @@
+export interface ApiUrls {
+  contacts         :  string;
+  events           :  string;
+  favorites        :  string;
+  folders          :  string;
+  groups           :  string;
+  home             :  string;
+  imageUrls        :  string;
+  reports          :  string;
+  search           :  string;
+  server           :  string;
+  sheets           :  string;
+  sights           :  string;
+  templates        :  string;
+  templatesPublic  :  string;
+  token            :  string;
+  users            :  string;
+  webhooks         :  string;
+  workspaces       :  string;
+}
