+++ lib/types/CreateClient.ts
@@ -1,0 +1,4 @@
+import {CreateClientOptions} from "./CreateClientOptions";
+import {SmartsheetClient} from "./SmartsheetClient";
+
+export type CreateClient = (options?: CreateClientOptions) => SmartsheetClient;
