+++ lib/types/CreateClientOptions.ts
@@ -1,0 +1,13 @@
+import {LoggerInstance} from "winston";
+
+export interface CreateClientOptions {
+  accessToken?: string;
+  userAgent?: string;
+  baseUrl?: string;
+  requestor?: any; // Custom HTTP client that will be used. TODO -> Evaluate if we want to keep this.
+  maxRetryDurationSeconds?: number;
+  calcRetryBackoff?: (retryCount: number, error?: any) => number;
+  logger?: LoggerInstance;
+  logLevel?: "error" | "warn" | "info" | "http" | "info" | "http" | "verbose" | "debug" | "silly";
+  loggerContainer?: any;
+}
