+++ lib/types/CreateOptions.ts
@@ -1,0 +1,5 @@
+import {ApiUrls} from "./ApiUrls";
+
+export interface CreateOptions {
+  apiUrls: ApiUrls;
+}
