+++ lib/types/SmartsheetClient.ts
@@ -1,0 +1,21 @@
+export interface SmartsheetClient {
+  constants: any;
+  contacts: any;
+  events: any;
+  favorites: any;
+  folders: any;
+  groups: any;
+  home: any;
+  images: any;
+  reports: any;
+  request: any;
+  search: any;
+  server: any;
+  sheets: any;
+  sights: any;
+  templates: any;
+  tokens: any;
+  users: any;
+  webhooks: any;
+  workspaces: any;
+}
