+++ lib/types/index.ts
@@ -1,0 +1,6 @@
+export * from "./SmartsheetClient";
+export * from "./CreateClientOptions";
+export * from "./ApiError";
+export * from "./CreateOptions";
+export * from "./ApiUrls";
+export * from "./CreateClient";
